backend/requirements.txt
Django==1.8.4
wsgiref==0.1.2
psycopg2==2.6.1
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
django-model-utils==2.3.1
drf-extensions==0.2.8
django-filter==0.11.0
